### 2025-07-24 #717 @ScottClaessens
coevolve: Fit Bayesian Generalized Dynamic Phylogenetic Models using 'Stan'
Submitting Author Name: Scott Claessens
Submitting Author Github Handle: @ScottClaessens
Other Package Authors Github handles: @ErikRingen
Repository:  https://github.com/ScottClaessens/coevolve
Version submitted: 0.1.0.9005
Submission type: Stats
Badge grade: gold
Due dates: TBD
Version accepted: TBD

---

```
Package: coevolve
Title: Fit Bayesian Generalized Dynamic Phylogenetic Models using 'Stan'
Version: 0.1.0.9005
    c(person("Scott", "Claessens", , "scott.claessens@gmail.com",
             role = c("aut", "cre"), comment = c(ORCID = "0000-0002-3562-6981")),
      person("Erik", "Ringen", , "erikjacob.ringen@uzh.ch",
             role = c("aut")))
Description: Fit Bayesian generalized dynamic phylogenetic models using 'Stan'.
    The package allows an abritrary number of variables with different response
    distributions to coevolve on a phylogenetic tree via an Ornstein-Uhlenbeck
    evolutionary process, allowing users to assess directionality between
    coevolving variables.
Roxygen: list(markdown = TRUE, roclets = c("namespace", "rd", "srr::srr_stats_roclet"))
    withr
    ape,
    bayesplot,
    cmdstanr,
    colorspace,
    gridExtra,
    Matrix,
    patchwork,
    phangorn,
    phaseR,
    phytools,
    posterior,
    tidyr
Additional_repositories: https://stan-dev.r-universe.dev/
    stan-dev/cmdstanr,
    mjg211/phaseR
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one or more appropriate boxes below):

- [x] A pre-submission inquiry has been approved in issue #715

The target audience is researchers in the life sciences and cultural evolution who would like to model how multiple variables have coevolved over evolutionary time. The coevolve package allows users to fit Bayesian generalised dynamic phylogenetic models in Stan. These models estimate the autoregressive and cross-lagged effects of variables along the branches of a phylogenetic tree. The current version of the package allows users to model different data types (e.g., binary, ordinal, continuous, count) while accounting for measurement error, missing data, and repeated measures.

This is the first implementation of a novel algorithm.

I don't think these guidelines apply to this package.

-    What grade of badge are you aiming for?

[gold](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-badges)

1. Compliance with a good number of standards beyond those identified as minimally necessary.
2. Demonstrating excellence in compliance with multiple standards from at least two broad sub-categories.
3. Internal aspects of package structure and design.

Currently, the `pkgcheck()` function is failing due to a [failing R CMD check workflow run on macOS](https://github.com/ScottClaessens/coevolve/actions/runs/16471617008). We are hoping that this is temporary failure due to GitHub runner caching.
