### 2025-11-13 a20ec1a
Add support for nutpie as an alternative sampler in coev_fit function
- Updated DESCRIPTION to suggest reticulate package.
- Modified NAMESPACE to include new S3 methods for nutpie_fit.
- Enhanced README with instructions for using nutpie.
- Updated coev_fit function to handle nutpie sampling, including argument checks and sampling logic.
- Added helper functions for nutpie integration, including checks for availability and model compilation.
- Implemented compatibility for plotting and extracting draws with nutpie.
- Added tests to ensure nutpie functionality works as expected.
