+++ R/nutpie_wrapper.R
@@ -1,0 +1,289 @@
+#' Create a wrapper object for nutpie fits that mimics cmdstanr interface
+#'
+#' @srrstats {G1.4a} Non-exported function documented here
+#'
+#' @description Creates an S3 object that wraps a nutpie trace and provides
+#'   methods compatible with cmdstanr's fit object interface. This allows
+#'   nutpie-fitted models to work with existing coevolve functions.
+#'
+#' @param trace Nutpie trace object (Python object via reticulate).
+#' @param draws_array A draws_array object from the posterior package.
+#' @param stan_variables Character vector of Stan variable names.
+#' @param iter_sampling Integer. Number of sampling iterations per chain.
+#' @param iter_warmup Integer. Number of warmup iterations per chain.
+#' @param chains Integer. Number of chains.
+#' @param seed Integer. Random seed used.
+#'
+#' @returns An object of class 'nutpie_fit' that mimics cmdstanr's interface.
+#'
+#' @noRd
+create_nutpie_wrapper <- function(trace, draws_array, stan_variables,
+                                  iter_sampling, iter_warmup, chains, seed = NULL) {
+  # Create wrapper object
+  wrapper <- list(
+    trace = trace,
+    draws_array = draws_array,
+    stan_variables = stan_variables,
+    iter_sampling = iter_sampling,
+    iter_warmup = iter_warmup,
+    chains = chains,
+    seed = seed
+  )
+  
+  class(wrapper) <- "nutpie_fit"
+  
+  # Add $draws() method for compatibility with cmdstanr interface
+  # This allows code like object$fit$draws() to work for both samplers
+  wrapper$draws <- function(variables = NULL, ...) {
+    draws.nutpie_fit(wrapper, variables = variables, ...)
+  }
+  
+  # Add $summary() method for compatibility with cmdstanr interface
+  # This allows code like object$fit$summary() to work for both samplers
+  wrapper$summary <- function(variables = NULL, ...) {
+    summary.nutpie_fit(wrapper, variables = variables, ...)
+  }
+  
+  # Add $metadata() method for compatibility with cmdstanr interface
+  # This allows code like object$fit$metadata() to work for both samplers
+  wrapper$metadata <- function() {
+    metadata.nutpie_fit(wrapper)
+  }
+  
+  # Add $num_chains() method for compatibility with cmdstanr interface
+  wrapper$num_chains <- function() {
+    return(wrapper$chains)
+  }
+  
+  # Add $diagnostic_summary() method for compatibility with cmdstanr interface
+  # nutpie stores divergence information in trace$sample_stats$diverging
+  wrapper$diagnostic_summary <- function(diagnostic = "divergences", quiet = TRUE) {
+    if (diagnostic == "divergences") {
+      # Extract divergence information from nutpie InferenceData
+      # Divergences are stored in trace$sample_stats$diverging
+      tryCatch({
+        if (reticulate::py_has_attr(wrapper$trace, "sample_stats")) {
+          sample_stats <- wrapper$trace$sample_stats
+          if (reticulate::py_has_attr(sample_stats, "data_vars")) {
+            # Check if diverging variable exists
+            py_builtins <- reticulate::import_builtins()
+            keys_obj <- sample_stats$data_vars$keys()
+            var_names <- reticulate::py_to_r(py_builtins$list(keys_obj))
+            
+            if ("diverging" %in% var_names) {
+              # Extract diverging array
+              diverging_xarray <- sample_stats[["diverging"]]
+              diverging_array <- diverging_xarray$values
+              diverging_r <- reticulate::py_to_r(diverging_array)
+              
+              # Count total number of divergences across all chains and draws
[…]
