+++ R/nutpie_helpers.R
@@ -1,0 +1,591 @@
+#' Check if nutpie is available via reticulate
+#'
+#' @srrstats {G1.4} Function is documented
+#'
+#' @description Checks if nutpie can be imported via reticulate. This function
+#'   attempts to import nutpie using the currently configured Python environment
+#'   (via reticulate). Users must configure reticulate to use a Python environment
+#'   with nutpie installed before using this function.
+#'
+#' @details To configure nutpie, you must:
+#' \enumerate{
+#'   \item Install nutpie in a Python environment: \code{pip install "nutpie[stan]"}
+#'   \item Configure reticulate to use that Python environment:
+#'     \itemize{
+#'       \item Set \code{RETICULATE_PYTHON} environment variable to the Python executable path
+#'       \item Or use \code{reticulate::use_python()} or \code{reticulate::use_virtualenv()}
+#'     }
+#' }
+#'
+#' @returns Logical. TRUE if nutpie is available, FALSE otherwise.
+#'
+#' @examples
+#' \dontrun{
+#' # Configure reticulate first (choose one):
+#' # Option 1: Set environment variable
+#' Sys.setenv(RETICULATE_PYTHON = "/path/to/python")
+#'
+#' # Option 2: Use Python directly
+#' reticulate::use_python("/path/to/python")
+#'
+#' # Option 3: Use virtual environment
+#' reticulate::use_virtualenv("~/.venvs/nutpie-env")
+#'
+#' # Then check availability
+#' if (check_nutpie_available()) {
+#'   # Use nutpie
+#' } else {
+#'   # Fall back to cmdstanr
+#' }
+#' }
+#'
+#' @export
+check_nutpie_available <- function() {
+  # Check if reticulate is available
+  if (!requireNamespace("reticulate", quietly = TRUE)) {
+    return(FALSE)
+  }
+  
+  # Try to import nutpie using the currently configured Python
+  # No automatic searching - user must configure reticulate explicitly
+  tryCatch({
+    nutpie <- reticulate::import("nutpie", convert = FALSE)
+    # If we get here, nutpie is available
+    return(TRUE)
+  }, error = function(e) {
+    # nutpie not available
+    return(FALSE)
+  })
+}
+
+#' Compile a Stan model using nutpie
+#'
+#' @srrstats {G1.4} Function is documented
+#'
+#' @description Compiles a Stan model string using nutpie's compile_stan_model
+#'   function. This is a wrapper around nutpie's Python API via reticulate.
+#'
+#' @param stan_code Character string containing Stan model code.
+#'
+#' @returns Compiled Stan model object (Python object via reticulate).
+#'
+#' @examples
+#' \dontrun{
+#' stan_code <- "
+#' data {
+#'   int<lower=0> N;
+#'   vector[N] y;
+#' }
+#' parameters {
+#'   real mu;
+#' }
+#' model {
+#'   mu ~ normal(0, 1);
+#'   y ~ normal(mu, 1);
+#' }
+#' "
+#' compiled <- nutpie_compile_stan_model(stan_code)
+#' }
+#'
+#' @export
+nutpie_compile_stan_model <- function(stan_code) {
+  # Check if nutpie is available
+  if (!check_nutpie_available()) {
+    stop2(
+      "nutpie is not available. Please install nutpie with: ",
+      "pip install 'nutpie[stan]'"
+    )
+  }
+  
+  # Import nutpie
+  nutpie <- reticulate::import("nutpie", convert = FALSE)
+  
+  # Compile Stan model
+  tryCatch({
+    compiled <- nutpie$compile_stan_model(code = stan_code)
+    return(compiled)
+  }, error = function(e) {
+    # Convert Python error to R error
[…]
