+++ man/summary.nutpie_fit.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/nutpie_wrapper.R
+\name{summary.nutpie_fit}
+\alias{summary.nutpie_fit}
+\title{Compute summary statistics for nutpie_fit object}
+\usage{
+\method{summary}{nutpie_fit}(object, variables = NULL, ...)
+}
+\arguments{
+\item{object}{A nutpie_fit object.}
+
+\item{variables}{Character vector of variable names to summarize. If NULL,
+summarizes all variables.}
+
+\item{...}{Named arguments where names are output column names and values
+are function names (strings) or formulas for computing statistics.}
+}
+\value{
+A data.frame with summary statistics, compatible with cmdstanr format.
+}
+\description{
+Computes summary statistics from a nutpie_fit object,
+compatible with cmdstanr's summary() method. This method mimics
+cmdstanr's interface where summary() is called with named arguments
+where names are column names and values are function names or formulas.
+}

[…]
