### 2025-11-14 c50d443
Enhance nutpie integration in coev_fit and related functions
- Updated coev_fit to support additional compilation and sampling arguments for nutpie.
- Make extract_samples function work with nutpie fits

- Updated documentation to reflect new parameters and usage instructions for nutpie features.
