+++ R/extract_samples.R
@@ -34,18 +34,47 @@ extract_samples <- function(object) {
 
 #' @export
 extract_samples.coevfit <- function(object) {
-  # get variables and rvars draws
+  # get variables and draws
   vars <- object$fit$metadata()$stan_variables
-  draws <- posterior::as_draws_rvars(object$fit$draws())
-  # reshape to 'rethinking' style list format
-  lapply(
-    vars,
-    \(var_name) {
-      posterior::draws_of(
-        draws[[var_name]],
-        with_chains = FALSE
-      )
+  draws_array <- object$fit$draws()
+  
+  # Convert to rvars format (groups indexed variables into arrays)
+  # e.g., A_diag[1], A_diag[2] -> A_diag array
+  draws_rvars <- posterior::as_draws_rvars(draws_array)
+  
+  # Extract each base variable (not individual indexed variables)
+  # rvars groups indexed variables, so we extract by base name
+  base_vars <- names(draws_rvars)
+  
+  samples_list <- lapply(base_vars, function(var_name) {
+    # Get the rvar object
+    var_rvar <- draws_rvars[[var_name]]
+    
+    # Extract draws as array (collapses chains)
+    # draws_of() returns arrays where first dimension is draws
+    var_draws <- posterior::draws_of(var_rvar, with_chains = FALSE)
+    
+    # Return as vector or array depending on dimensions
+    # For scalar variables: return as vector
+    # For array variables: return as array/matrix
+    if (is.array(var_draws)) {
+      n_dims <- length(dim(var_draws))
+      if (n_dims == 1) {
+        # 1D array -> vector
+        return(as.vector(var_draws))
+      } else {
+        # Multi-dimensional array -> keep as array
+        # First dimension is draws, remaining dimensions are parameter dimensions
+        return(var_draws)
+      }
+    } else {
+      # Not an array -> return as vector
+      return(as.vector(var_draws))
     }
-  ) |>
-    stats::setNames(vars)
+  })
+  
+  # Set names
+  names(samples_list) <- base_vars
+  
+  return(samples_list)
 }
