+++ R/nutpie_helpers.R
@@ -4,16 +4,19 @@
 #'
 #' @description Checks if nutpie can be imported via reticulate. This function
 #'   attempts to import nutpie using the currently configured Python environment
-#'   (via reticulate). Users must configure reticulate to use a Python environment
-#'   with nutpie installed before using this function.
+#'   (via reticulate). Users must configure reticulate to use a Python
+#'   environment with nutpie installed before using this function.
 #'
 #' @details To configure nutpie, you must:
 #' \enumerate{
-#'   \item Install nutpie in a Python environment: \code{pip install "nutpie[stan]"}
+#'   \item Install nutpie in a Python environment: \code{pip install
+#'     "nutpie[stan]"}
 #'   \item Configure reticulate to use that Python environment:
 #'     \itemize{
-#'       \item Set \code{RETICULATE_PYTHON} environment variable to the Python executable path
-#'       \item Or use \code{reticulate::use_python()} or \code{reticulate::use_virtualenv()}
+#'       \item Set \code{RETICULATE_PYTHON} environment variable to the Python
+#'         executable path
+#'       \item Or use \code{reticulate::use_python()} or
+#'         \code{reticulate::use_virtualenv()}
 #'     }
 #' }
 #'
@@ -41,20 +44,19 @@
 #'
 #' @export
 check_nutpie_available <- function() {
-  # Check if reticulate is available
+  # check if reticulate is available
   if (!requireNamespace("reticulate", quietly = TRUE)) {
     return(FALSE)
   }
-  
-  # Try to import nutpie using the currently configured Python
-  # No automatic searching - user must configure reticulate explicitly
+  # try to import nutpie using the currently configured Python
+  # no automatic searching - user must configure reticulate explicitly
   tryCatch({
-    nutpie <- reticulate::import("nutpie", convert = FALSE)
+    reticulate::import("nutpie", convert = FALSE)
     # If we get here, nutpie is available
-    return(TRUE)
+    TRUE
   }, error = function(e) {
     # nutpie not available
-    return(FALSE)
+    FALSE
   })
 }
 
@@ -69,17 +71,21 @@ check_nutpie_available <- function() {
 #' @param ... Additional arguments passed to \code{nutpie.compile_stan_model()}.
 #'   Valid parameters include:
 #'   \itemize{
-#'     \item \code{extra_stanc_args}: List of strings. Arguments passed to stanc3
-#'       (Stan compiler). Common options include \code{"--O0"} (no optimization),
-#'       \code{"--O1"} (optimization level 1). Example: \code{extra_stanc_args = list("--O1")}.
-#'     \item \code{extra_compile_args}: List of strings. Arguments passed to Make
-#'       (C++ compiler). Common options include \code{"STAN_THREADS=true"} to enable
-#'       threading. Example: \code{extra_compile_args = list("STAN_THREADS=true")}.
-#'     \item \code{filename}: Character string. Path to Stan file (alternative to \code{stan_code}).
+#'     \item \code{extra_stanc_args}: List of strings. Arguments passed to
+#'       stanc3 (Stan compiler). Common options include \code{"--O0"} (no
+#'       optimization), \code{"--O1"} (optimization level 1). Example:
+#'       \code{extra_stanc_args = list("--O1")}.
+#'     \item \code{extra_compile_args}: List of strings. Arguments passed to
+#'       Make (C++ compiler). Common options include \code{"STAN_THREADS=true"}
+#'       to enable threading. Example: \code{extra_compile_args =
+#'       list("STAN_THREADS=true")}.
+#'     \item \code{filename}: Character string. Path to Stan file (alternative
+#'       to \code{stan_code}).
 #'     \item \code{dims}: Named list. Dimension information.
 #'     \item \code{coords}: Named list. Coordinate information.
[…]
