### 2025-11-15 efcb253
- Added caching for unique branch lengths and their mappings to improve efficiency in model computations. - Updated data structures in `coev_make_standata` to include unique lengths and their indices. - Improved handling of drift calculations by integrating cached matrices for unique lengths.
