+++ R/coev_make_standata.R
@@ -334,6 +334,37 @@ coev_make_standata <- function(data, variables, id, tree,
     stan_ts[t, ] <- parent_time
     stan_tip[t, ] <- tip
   }
+  # identify unique branch lengths for caching
+  all_branch_lengths <- as.vector(stan_ts[stan_ts != -99 & stan_ts > 0])
+  unique_lengths <- sort(unique(all_branch_lengths))
+  n_unique_lengths <- length(unique_lengths)
+  
+  # create mapping from segments to unique length indices (0 for root, 1+ for branches)
+  stan_length_index <- matrix(0L, n_tree, n_seg)
+  for (t in seq_along(tree)) {
+    for (i in 2:n_seg) {
+      if (stan_ts[t, i] != -99 && stan_ts[t, i] > 0) {
+        idx <- match(stan_ts[t, i], unique_lengths)
+        if (!is.na(idx)) {
+          stan_length_index[t, i] <- idx
+        }
+      }
+    }
+  }
+  
+  # create mapping from tips to segments (for VCV_tips caching)
+  stan_tip_to_seg <- matrix(0L, n_tree, length(tree[[1]]$tip.label))
+  for (t in seq_along(tree)) {
+    for (seg in 1:n_seg) {
+      if (stan_tip[t, seg] == 1) {
+        tip_node <- stan_node_seq[t, seg]
+        if (tip_node >= 1 && tip_node <= length(tree[[t]]$tip.label)) {
+          stan_tip_to_seg[t, tip_node] <- seg
+        }
+      }
+    }
+  }
+  
   # get data matrix
   y <- list()
   for (j in seq_along(variables)) {
@@ -394,7 +425,7 @@ coev_make_standata <- function(data, variables, id, tree,
   # data list for stan
   sd <- list(
     N_tips = length(tree[[1]]$tip.label), # number of tips
-    N_tree = n_tree,                      # number of tips
+    N_tree = n_tree,                      # number of trees
     N_obs = nrow(y),                      # number of observations
     J = length(variables),                # number of variables
     N_seg = n_seg,                        # number of segments in the tree
@@ -406,7 +437,11 @@ coev_make_standata <- function(data, variables, id, tree,
     num_effects = sum(effects_mat),       # number of effects being estimated
     y = y,                                # observed data
     miss = miss,                          # are data points missing?
-    tip_id = tip_id                       # tip ids
+    tip_id = tip_id,                      # tip ids
+    N_unique_lengths = n_unique_lengths,  # number of unique branch lengths
+    unique_lengths = unique_lengths,      # unique branch lengths for caching
+    length_index = stan_length_index,     # mapping from segments to unique lengths
+    tip_to_seg = stan_tip_to_seg         # mapping from tips to segments
   )
   # add distance matrix if specified
   if (!is.null(dist_mat)) sd[["dist_mat"]] <- dist_mat
