+++ R/coev_make_stancode.R
@@ -467,7 +467,11 @@ write_data_block <- function(measurement_error, dist_mat) {
       !is.null(measurement_error),
       "  matrix[N_obs,J] se; // squared standard errors\n", ""
     ),
-    "  array[N_obs] int<lower=1> tip_id; // group index between 1 and N_tips\n"
+    "  array[N_obs] int<lower=1> tip_id; // group index between 1 and N_tips\n",
+    "  int<lower=1> N_unique_lengths; // number of unique branch lengths\n",
+    "  array[N_unique_lengths] real unique_lengths; // unique branch lengths for caching\n",
+    "  array[N_tree, N_seg] int<lower=0> length_index; // mapping from segments to unique lengths\n",
+    "  array[N_tree, N_tips] int<lower=0> tip_to_seg; // mapping from tips to segments\n"
   )
   # add distance matrix if user has defined one
   if (!is.null(dist_mat)) {
@@ -695,48 +699,84 @@ write_transformed_pars_block <- function(data, distributions, id, dist_mat,
     "  }\n",
     "  // calculate asymptotic covariance\n",
     "  Q_inf = ksolve(A, Q);\n",
-    "  // loop over phylogenetic trees\n",
-    "  for (t in 1:N_tree) {\n",
+    "  array[N_unique_lengths] matrix[J,J] L_VCV_tips_cache;\n",
+    "  {\n",
+    "    array[N_unique_lengths] matrix[J,J] A_delta_cache;\n",
+    "    array[N_unique_lengths] matrix[J,J] VCV_cache;\n",
+    "    array[N_unique_lengths] matrix[J,J] L_VCV_cache;\n",
+    "    array[N_unique_lengths] matrix[J,J] A_solve_cache;\n",
+    "    for (u in 1:N_unique_lengths) {\n",
+    "      A_delta_cache[u] = matrix_exp(A * unique_lengths[u]);\n",
+    "      VCV_cache[u] = Q_inf - quad_form_sym(Q_inf, A_delta_cache[u]');\n",
+    "      L_VCV_cache[u] = cholesky_decompose(VCV_cache[u]);\n",
+      "      A_solve_cache[u] = A \\ add_diag(A_delta_cache[u], -1);\n",
+    "      for (i in 1:J) {\n",
+    "        for (j in 1:i) {\n",
+    "          real val = 0.5 * (A_solve_cache[u][i, j] + A_solve_cache[u][j, i]);\n",
+    "          A_solve_cache[u][i, j] = val;\n",
+    "          A_solve_cache[u][j, i] = val;\n",
+    "        }\n",
+    "      }\n",
+    "    }\n",
+    "    L_VCV_tips_cache = L_VCV_cache;\n",
+    "    for (t in 1:N_tree) {\n",
     "    // setting ancestral states and placeholders\n",
     "    eta[t, node_seq[t, 1]] = eta_anc[t];\n",
     "    VCV_tips[t, node_seq[t, 1]] = diag_matrix(rep_vector(-99, J));\n",
     "    for (i in 2:N_seg) {\n",
-    "      matrix[J,J] A_delta; // amount of deterministic change\n",
-    "      matrix[J,J] VCV; // vcov matrix of stochastic change\n",
-    "      vector[J] drift_seg; // accumulated drift over the segment\n",
-    "      A_delta = matrix_exp(A * ts[t, i]);\n",
-    "      VCV = Q_inf - quad_form_sym(Q_inf, A_delta');\n",
-    "      drift_seg = cholesky_decompose(VCV) * z_drift[t, i-1];\n",
+    "      matrix[J,J] A_delta;\n",
+    "      matrix[J,J] VCV;\n",
[…]
