### 2025-11-15 e5fd056
Enhance transformed parameters block to include measurement error handling
- Updated `write_transformed_pars_block` to accept `measurement_error` parameter.
- Implemented caching of residual covariance base for improved efficiency in model computations.
- Adjusted model and quantities blocks to utilize the new residual covariance base.
