### 2025-11-15 07ebacb
Refactor residual covariance handling in transformed parameters block
