### 2025-11-15 713eb15
Refactor transformed parameters block to improve covariance handling
- Removed unused L_VCV_cache and replaced it with L_VCV_tips_cache for clarity.
- Introduced A_solve_b_cache to streamline calculations involving the A_solve matrix.
- Updated references in the model to utilize the new A_solve_b for enhanced efficiency.
