+++ R/coev_make_stancode.R
@@ -716,12 +716,12 @@ write_transformed_pars_block <- function(data, distributions, id, dist_mat,
     "  {\n",
     "    array[N_unique_lengths] matrix[J,J] A_delta_cache;\n",
     "    array[N_unique_lengths] matrix[J,J] VCV_cache;\n",
-    "    array[N_unique_lengths] matrix[J,J] L_VCV_cache;\n",
     "    array[N_unique_lengths] matrix[J,J] A_solve_cache;\n",
+    "    array[N_unique_lengths] vector[J] A_solve_b_cache;\n",
     "    for (u in 1:N_unique_lengths) {\n",
     "      A_delta_cache[u] = matrix_exp(A * unique_lengths[u]);\n",
     "      VCV_cache[u] = Q_inf - quad_form_sym(Q_inf, A_delta_cache[u]');\n",
-    "      L_VCV_cache[u] = cholesky_decompose(VCV_cache[u]);\n",
+    "      L_VCV_tips_cache[u] = cholesky_decompose(VCV_cache[u]);\n",
       "      A_solve_cache[u] = A \\ add_diag(A_delta_cache[u], -1);\n",
     "      for (i in 1:J) {\n",
     "        for (j in 1:i) {\n",
@@ -730,8 +730,8 @@ write_transformed_pars_block <- function(data, distributions, id, dist_mat,
     "          A_solve_cache[u][j, i] = val;\n",
     "        }\n",
     "      }\n",
+      "      A_solve_b_cache[u] = A_solve_cache[u] * b;\n",
     "    }\n",
-    "    L_VCV_tips_cache = L_VCV_cache;\n",
     "    for (t in 1:N_tree) {\n",
     "    // setting ancestral states and placeholders\n",
     "    eta[t, node_seq[t, 1]] = eta_anc[t];\n",
@@ -741,30 +741,30 @@ write_transformed_pars_block <- function(data, distributions, id, dist_mat,
     "      matrix[J,J] VCV;\n",
     "      vector[J] drift_seg;\n",
     "      matrix[J,J] L_VCV;\n",
-    "      matrix[J,J] A_solve;\n",
-    "      if (length_index[t, i] > 0) {\n",
-    "        A_delta = A_delta_cache[length_index[t, i]];\n",
-    "        VCV = VCV_cache[length_index[t, i]];\n",
-    "        L_VCV = L_VCV_cache[length_index[t, i]];\n",
-    "        A_solve = A_solve_cache[length_index[t, i]];\n",
+    "      vector[J] A_solve_b;\n",
+      "      if (length_index[t, i] > 0) {\n",
+      "        A_delta = A_delta_cache[length_index[t, i]];\n",
+      "        VCV = VCV_cache[length_index[t, i]];\n",
+      "        L_VCV = L_VCV_tips_cache[length_index[t, i]];\n",
+      "        A_solve_b = A_solve_b_cache[length_index[t, i]];\n",
     "      } else {\n",
     "        A_delta = matrix_exp(A * ts[t, i]);\n",
     "        VCV = Q_inf - quad_form_sym(Q_inf, A_delta');\n",
     "        L_VCV = cholesky_decompose(VCV);\n",
-    "        A_solve = A \\ add_diag(A_delta, -1);\n",
+    "        A_solve_b = (A \\ add_diag(A_delta, -1)) * b;\n",
     "      }\n",
     "      drift_seg = L_VCV * z_drift[t, i-1];\n",
     "      // if not a tip, add the drift parameter\n",
     "      if (tip[t, i] == 0) {\n",
     "        eta[t, node_seq[t, i]] = to_vector(\n",
[…]
