### 2025-11-15 dc2c643
Removed measurement error handling from `write_transformed_pars_block` for clarity. - Updated references to covariance matrices to improve efficiency in model computations. - Adjusted handling of `A_solve` and `L_VCV` to enhance performance and maintain consistency.
