+++ R/coev_make_stancode.R
@@ -276,7 +276,7 @@ coev_make_stancode <- function(data, variables, id, tree,
     "\n",
     write_transformed_pars_block(data, distributions, id, dist_mat,
                                  dist_cov, estimate_correlated_drift,
-                                 estimate_residual, measurement_error),
+                                 estimate_residual),
     "\n",
     write_model_block(data, distributions, id, dist_mat, priors,
                       measurement_error, estimate_correlated_drift,
@@ -636,7 +636,7 @@ write_parameters_block <- function(data, variables, distributions, id, dist_mat,
 #' @noRd
 write_transformed_pars_block <- function(data, distributions, id, dist_mat,
                                          dist_cov, estimate_correlated_drift,
-                                         estimate_residual, measurement_error) {
+                                         estimate_residual) {
   sc_transformed_parameters <- paste0(
     "transformed parameters{\n",
     "  array[N_tree, N_seg] vector[J] eta;\n",
@@ -699,29 +699,16 @@ write_transformed_pars_block <- function(data, distributions, id, dist_mat,
     "  }\n",
     "  // calculate asymptotic covariance\n",
     "  Q_inf = ksolve(A, Q);\n",
-    ifelse(
-      estimate_residual && any(duplicated(data[, id])),
-      paste0(
-        "  // cache residual covariance components (computed once per iteration)\n",
-        ifelse(
-          !is.null(measurement_error),
-          "  matrix[J,J] residual_cov_base = quad_form_diag(L_residual * L_residual', sigma_residual);\n",
-          ""
-        ),
-        "  matrix[J,J] L_residual_scaled = diag_pre_multiply(sigma_residual, L_residual);\n"
-      ),
-      ""
-    ),
     "  array[N_unique_lengths] matrix[J,J] L_VCV_tips_cache;\n",
     "  {\n",
     "    array[N_unique_lengths] matrix[J,J] A_delta_cache;\n",
     "    array[N_unique_lengths] matrix[J,J] VCV_cache;\n",
+    "    array[N_unique_lengths] matrix[J,J] L_VCV_cache;\n",
     "    array[N_unique_lengths] matrix[J,J] A_solve_cache;\n",
-    "    array[N_unique_lengths] vector[J] A_solve_b_cache;\n",
     "    for (u in 1:N_unique_lengths) {\n",
     "      A_delta_cache[u] = matrix_exp(A * unique_lengths[u]);\n",
     "      VCV_cache[u] = Q_inf - quad_form_sym(Q_inf, A_delta_cache[u]');\n",
-    "      L_VCV_tips_cache[u] = cholesky_decompose(VCV_cache[u]);\n",
+    "      L_VCV_cache[u] = cholesky_decompose(VCV_cache[u]);\n",
       "      A_solve_cache[u] = A \\ add_diag(A_delta_cache[u], -1);\n",
     "      for (i in 1:J) {\n",
     "        for (j in 1:i) {\n",
@@ -729,9 +716,9 @@ write_transformed_pars_block <- function(data, distributions, id, dist_mat,
     "          A_solve_cache[u][i, j] = val;\n",
     "          A_solve_cache[u][j, i] = val;\n",
     "        }\n",
-    "      }\n",
-      "      A_solve_b_cache[u] = A_solve_cache[u] * b;\n",
+      "      }\n",
     "    }\n",
+    "    L_VCV_tips_cache = L_VCV_cache;\n",
     "    for (t in 1:N_tree) {\n",
     "    // setting ancestral states and placeholders\n",
     "    eta[t, node_seq[t, 1]] = eta_anc[t];\n",
[…]
