### 2025-11-16 065d39d
More robust variable conversion and error handling in convert_nutpie_draws function
