+++ R/nutpie_helpers.R
@@ -42,7 +42,7 @@
 #' }
 #' }
 #'
-#' @export
+#' @keywords internal
 check_nutpie_available <- function() {
   # check if reticulate is available
   if (!requireNamespace("reticulate", quietly = TRUE)) {
@@ -121,7 +121,7 @@ check_nutpie_available <- function() {
 #' )
 #' }
 #'
-#' @export
+#' @keywords internal
 nutpie_compile_stan_model <- function(stan_code, ...) {
   # check if nutpie is available
   if (!check_nutpie_available()) {
@@ -307,7 +307,7 @@ convert_r_to_python_data <- function(data_list) {
 #'                       seed = 12345L)
 #' }
 #'
-#' @export
+#' @keywords internal
 nutpie_sample <- function(stan_code, data_list,
                           num_chains = 4L,
                           num_samples = 1000L,
@@ -451,7 +451,7 @@ def nutpie_call_with_data(compiled_model, data_dict):
 #' draws <- convert_nutpie_draws(trace)
 #' }
 #'
-#' @export
+#' @keywords internal
 convert_nutpie_draws <- function(trace) {
   # check if posterior package is available
   if (!requireNamespace("posterior", quietly = TRUE)) {
