+++ man/nutpie_compile_stan_model.Rd
@@ -68,3 +68,4 @@ compiled_threads <- nutpie_compile_stan_model(
 }
 
 }
+\keyword{internal}
