+++ man/check_nutpie_available.Rd
@@ -50,3 +50,4 @@ if (check_nutpie_available()) {
 }
 
 }
+\keyword{internal}
