+++ NAMESPACE
@@ -10,7 +10,6 @@ S3method(stancode,coevfit)
 S3method(standata,coevfit)
 S3method(summary,coevfit)
 S3method(summary,nutpie_fit)
-export(check_nutpie_available)
 export(coev_calculate_delta_theta)
 export(coev_calculate_theta)
 export(coev_fit)
@@ -24,12 +23,7 @@ export(coev_plot_selection_gradient)
 export(coev_plot_trait_values)
 export(coev_pred_series)
 export(coev_simulate_coevolution)
-export(convert_nutpie_draws)
-export(draws)
 export(extract_samples)
-export(metadata)
-export(nutpie_compile_stan_model)
-export(nutpie_sample)
 export(save_coevfit)
 export(stancode)
 export(standata)
