### 2025-11-17 ab79c3c
Use noRd tag for internal nutpie functions
