### 2025-11-17 da47b45
Change 'sampler' argument to 'backend'
