### 2025-11-17 f3e9702
nutpie instlation in CI
