+++ .github/workflows/R-CMD-check.yaml
@@ -69,11 +69,35 @@ jobs:
             rcmdcheck
             readr
             rlang
+            reticulate
             rmarkdown
             stringr
             testthat
             tidyr
             withr
+
+      - name: Set up Python for nutpie
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+
+      - name: Install nutpie
+        run: |
+          python -m pip install --upgrade pip
+          pip install "nutpie[stan]"
+        shell: bash
+
+      - name: Configure reticulate for nutpie
+        run: |
+          # Use python from PATH (setup-python ensures it's available)
+          python_path=$(which python || which python3)
+          if [ -z "$python_path" ]; then
+            echo "Error: Python not found in PATH"
+            exit 1
+          fi
+          echo "RETICULATE_PYTHON=$python_path" >> $GITHUB_ENV
+          echo "Configured RETICULATE_PYTHON=$python_path"
+        shell: bash
 
       - name: Build Cmdstan
         run: |
