### 2025-11-17 5f547f5
- Refined argument mapping for cmdstanr to nutpie in coev_fit, ensuring essential parameters are correctly handled. - Added checks to convert 'parallel_chains' to 'cores' for nutpie compatibility. - Updated tests to verify nutpie functionality - Try to get Windows CI working
