+++ .github/workflows/R-CMD-check.yaml
@@ -89,12 +89,17 @@ jobs:
 
       - name: Configure reticulate for nutpie
         run: |
-          # Use python from PATH (setup-python ensures it's available)
-          python_path=$(which python || which python3)
+          # Use Python to get its own executable path (works cross-platform)
+          # This is more reliable than 'which' on Windows
+          python_path=$(python -c "import sys; print(sys.executable)")
           if [ -z "$python_path" ]; then
-            echo "Error: Python not found in PATH"
+            echo "Error: Could not determine Python path"
+            python --version || echo "Python command not available"
             exit 1
           fi
+          # On Windows, convert backslashes to forward slashes for consistency
+          # (reticulate can handle both, but forward slashes are more portable)
+          python_path=$(echo "$python_path" | sed 's|\\|/|g')
           echo "RETICULATE_PYTHON=$python_path" >> $GITHUB_ENV
           echo "Configured RETICULATE_PYTHON=$python_path"
         shell: bash
