### 2025-11-17 0f0b678
Add tests of package functions with nutpie models
