### 2025-11-17 d1abb3c
Additional tests for coev_fit()
