### 2025-11-17 60dbb91
Refactor error handling in convert_nutpie_draws to simplify array conversion logic Added tests for nutpie_compile_stan_model and nutpie_sample to verify compilation arguments and ensure correct handling of function parameters.
