### 2025-11-17 cc7bc37
optimization option for nutpie_compile_stan_model in documentation
