+++ R/nutpie_helpers.R
@@ -119,7 +119,7 @@ check_nutpie_available <- function() {
 #'   stan_code,
 #'   extra_compile_args = list("STAN_THREADS=true")
 #' )
-#' 
+#'
 #' # compile with CPU architecture optimization
 #' compiled_optimized <- nutpie_compile_stan_model(
 #'   stan_code,
