### 2025-11-18 7239e95
Minor changes and additional tests
