### 2025-11-18 0ecc47a
Remove low_rank_modified_mass_matrix parameter from coev_fit function and documentation
