### 2025-11-18 1737f1e
Fix failing tests
