+++ tests/testthat/test-coev_make_standata.R
@@ -803,7 +803,9 @@ test_that("coev_make_standata() returns a list with correct names for Stan", {
   expect_equal(
     names(sd1),
     c("N_tips", "N_tree", "N_obs", "J", "N_seg", "node_seq", "parent", "ts",
-      "tip", "effects_mat", "num_effects", "y", "miss", "tip_id", "prior_only")
+      "tip", "effects_mat", "num_effects", "y", "miss", "tip_id",
+      "N_unique_lengths", "unique_lengths", "length_index", "tip_to_seg",
+      "prior_only")
   )
   expect_equal(sd1$prior_only, 0)
   # include distance matrix
@@ -828,8 +830,9 @@ test_that("coev_make_standata() returns a list with correct names for Stan", {
   expect_equal(
     names(sd2),
     c("N_tips", "N_tree", "N_obs", "J", "N_seg", "node_seq", "parent", "ts",
-      "tip", "effects_mat", "num_effects", "y", "miss", "tip_id", "dist_mat",
-      "prior_only")
+      "tip", "effects_mat", "num_effects", "y", "miss", "tip_id",
+      "N_unique_lengths", "unique_lengths", "length_index", "tip_to_seg",
+      "dist_mat", "prior_only")
   )
   expect_equal(sd2$prior_only, 0)
   # set prior only
@@ -849,7 +852,9 @@ test_that("coev_make_standata() returns a list with correct names for Stan", {
   expect_equal(
     names(sd3),
     c("N_tips", "N_tree", "N_obs", "J", "N_seg", "node_seq", "parent", "ts",
-      "tip", "effects_mat", "num_effects", "y", "miss", "tip_id", "prior_only")
+      "tip", "effects_mat", "num_effects", "y", "miss", "tip_id",
+      "N_unique_lengths", "unique_lengths", "length_index", "tip_to_seg",
+      "prior_only")
   )
   expect_equal(sd3$prior_only, 1)
 })
@@ -1079,7 +1084,9 @@ test_that("coev_make_standata() works with tibbles", {
   expect_equal(
     names(sd),
     c("N_tips", "N_tree", "N_obs", "J", "N_seg", "node_seq", "parent", "ts",
-      "tip", "effects_mat", "num_effects", "y", "miss", "tip_id", "prior_only")
+      "tip", "effects_mat", "num_effects", "y", "miss", "tip_id",
+      "N_unique_lengths", "unique_lengths", "length_index", "tip_to_seg",
+      "prior_only")
   )
   expect_equal(sd$prior_only, 0)
 })
