### 2025-11-18 b7c06e0
Indentation fix in coev_make_stancode()
