+++ R/coev_make_stancode.R
@@ -721,10 +721,8 @@ write_transformed_pars_block <- function(data, distributions, id, dist_mat,
     "      A_solve_cache[u] = A \\ add_diag(A_delta_cache[u], -1);\n",
     "      for (i in 1:J) {\n",
     "        for (j in 1:i) {\n",
-    paste0(
-      "          real val = 0.5 * (A_solve_cache[u][i, j] + ",
-      "A_solve_cache[u][j, i]);\n"
-    ),
+    "          real val = 0.5 * (A_solve_cache[u][i, j] + ",
+    "A_solve_cache[u][j, i]);\n",
     "          A_solve_cache[u][i, j] = val;\n",
     "          A_solve_cache[u][j, i] = val;\n",
     "        }\n",
@@ -732,42 +730,43 @@ write_transformed_pars_block <- function(data, distributions, id, dist_mat,
     "    }\n",
     "    L_VCV_tips_cache = L_VCV_cache;\n",
     "    for (t in 1:N_tree) {\n",
-    "    // setting ancestral states and placeholders\n",
-    "    eta[t, node_seq[t, 1]] = eta_anc[t];\n",
-    "    VCV_tips[t, node_seq[t, 1]] = diag_matrix(rep_vector(-99, J));\n",
-    "    for (i in 2:N_seg) {\n",
-    "      matrix[J,J] A_delta;\n",
-    "      matrix[J,J] VCV;\n",
-    "      vector[J] drift_seg;\n",
-    "      matrix[J,J] L_VCV;\n",
-    "      matrix[J,J] A_solve;\n",
-    "      if (length_index[t, i] > 0) {\n",
-    "        A_delta = A_delta_cache[length_index[t, i]];\n",
-    "        VCV = VCV_cache[length_index[t, i]];\n",
-    "        L_VCV = L_VCV_cache[length_index[t, i]];\n",
-    "        A_solve = A_solve_cache[length_index[t, i]];\n",
-    "      } else {\n",
-    "        A_delta = matrix_exp(A * ts[t, i]);\n",
-    "        VCV = Q_inf - quad_form_sym(Q_inf, A_delta');\n",
-    "        L_VCV = cholesky_decompose(VCV);\n",
-    "        A_solve = A \\ add_diag(A_delta, -1);\n",
+    "      // setting ancestral states and placeholders\n",
+    "      eta[t, node_seq[t, 1]] = eta_anc[t];\n",
+    "      VCV_tips[t, node_seq[t, 1]] = diag_matrix(rep_vector(-99, J));\n",
+    "      for (i in 2:N_seg) {\n",
+    "        matrix[J,J] A_delta;\n",
+    "        matrix[J,J] VCV;\n",
+    "        vector[J] drift_seg;\n",
+    "        matrix[J,J] L_VCV;\n",
+    "        matrix[J,J] A_solve;\n",
+    "        if (length_index[t, i] > 0) {\n",
+    "          A_delta = A_delta_cache[length_index[t, i]];\n",
+    "          VCV = VCV_cache[length_index[t, i]];\n",
+    "          L_VCV = L_VCV_cache[length_index[t, i]];\n",
+    "          A_solve = A_solve_cache[length_index[t, i]];\n",
+    "        } else {\n",
+    "          A_delta = matrix_exp(A * ts[t, i]);\n",
+    "          VCV = Q_inf - quad_form_sym(Q_inf, A_delta');\n",
+    "          L_VCV = cholesky_decompose(VCV);\n",
+    "          A_solve = A \\ add_diag(A_delta, -1);\n",
+    "        }\n",
+    "        drift_seg = L_VCV * z_drift[t, i-1];\n",
+    "        // if not a tip, add the drift parameter\n",
[…]
