### 2025-11-22 9d9687b
Add tests for caching computations
