### 2025-11-22 7ea24bf
Update NEWS.md
