### 2025-11-24 f799c8f
Add L_VCV_tips matrix and update related calculations in coev_make_stancode.R
