### 2025-12-06 90a3bd9
Fix issue with summary when estimate_residual = FALSE #95
