### 2025-12-06 cda231d
Update NEWS
