+++ R/summary.R
@@ -196,7 +196,7 @@ summary.coevfit <- function(object, prob = 0.95, robust = FALSE, ...) {
   # summarise residual sds and correlations
   sd_residual <- NULL
   cor_residual <- NULL
-  if (any(duplicated(object$data[, object$id])) & object$estimate_residual) {
+  if (any(duplicated(object$data[, object$id])) && object$estimate_residual) {
     # sd parameters
     sd_residual <- s[stringr::str_starts(s$variable, "sigma_residual"), ]
     rownames(sd_residual) <- paste0(
@@ -269,7 +269,7 @@ summary.coevfit <- function(object, prob = 0.95, robust = FALSE, ...) {
       rhats          = object$fit$summary(NULL, "rhat")$rhat
     )
   class(out) <- "coevsummary"
-  return(out)
+  out
 }
 
 #' Print a summary of a fitted model represented by a \code{coevfit} object
