### 2026-02-06 #717 @ScottClaessens
@mpadge @natydasilva @ldecicco-USGS Just checking in again about this package review. Has there been any progress?

Update from our end: our methods paper that uses this package was under review at Methods Ecology and Evolution, and we recently got a minor revision on the paper due at the end of Feb. Hopefully the paper will be accepted soon after.

It would be great to get feedback on the package before the paper is out, just so we can iron out any kinks before people start to (hopefully) use the package more widely.
