### 2026-04-16 #717 @emilyriederer
Hi all -- I'm currently the rotating EiC and checking in on quiet threads. Bumping for @natydasilva 's visibility. Thank you!
