### 2026-04-16 #717 @ScottClaessens
Thanks @emilyriederer! Just another update on this. Our methods paper introducing this package was recently accepted at Methods in Ecology and Evolution. The paper is now in press.
