### 2026-05-03 #717 @maurolepore
Congratulations on the MEE acceptance! 🥳

I see the next step is editor checks by @natydasilva

Thanks for your patience and for sharing your work with the rOpenSci community.
