### 2026-05-13 #717 @maurolepore
Hi @ScottClaessens and @mpadge

I reached out to @natydasilva on slack to discuss her availability. At the moment she lacks the band-with require to continue handling this submission.

I'll share the news with our editors board so we can find a replacement ASAP.

We apologize for the inconvenience and again thank you for your patience.
