### 2026-08-16 #717 @krzbar
## Package Review

- Briefly describe any working relationship you may have (had) with the package authors (or otherwise remove this statement)

---

### Compliance with Standards

- [X] This package complies with a sufficient number of standards for a (bronze/silver/gold) badge
- [ ] This grade of badge is the same as what the authors wanted to achieve

The following standards currently deemed non-applicable (through tags of `@srrstatsNA`) could potentially be applied to future versions of this software: (Please specify)

Please also comment on any standards which you consider either particularly well, or insufficiently, documented.

For packages aiming for silver or gold badges:

- [x] This package extends beyond minimal compliance with standards in the following ways: (please describe)
-- (positive) The package is applicable to at least two fields: phylogenetic comparative methods, and phylogenetic linguistics.
-- (positive) Multipe ways to do model fitting.
-- (positive) roxygen2 is used, and most non-exported functions are documented (but not all, e.g., normalize_nuts_sampler()).
-- (positive) Code to reproduce results in accompanying articles is deposited in Zenodo and GitHub.
-- (negative, main reason for silver) I cannot find in the package's documentation what the underlying stochastis model actually is. Only in coev_simulate_coevolution.Rd is VAR(1) model mentioned, but this constradicts the SDE description and makes it
unclear how the phylogeny enters the model.
-- (negative) The example for setting priors in README.Rmd is very limited, and this reflects the point above, if the model is unclear, then setting priors is unclear? Are, e.g., multivariate priors possible?
---

### General Review

#### Documentation

- [ X ] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
- [ X ] **Installation instructions:** for the development version of package and any non-standard dependencies in README
- [ N/A ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING
there is a GitHub page and Author contact details in DESCRIPTION
- [ X ] The documentation is sufficient to enable general use of the package beyond one specific use case

The following sections of this template include questions intended to be used as guides to provide general, descriptive responses. Please remove this, and any subsequent lines that are not relevant or necessary for your final review.

#### Algorithms

- How well are algorithms encoded?
N/A code is pleasantly written, but all calculations are delegated to STAN.
- Is the choice of computer language appropriate for that algorithm, and/or envisioned use of package?
Yes.
- Are aspects of algorithmic scaling sufficiently documented and tested?
No, I cannot see such discussion in the vignettes.
- Are there any aspects of algorithmic implementation which could be improved?

#### Testing

- Regardless of actual coverage of tests, are there any fundamental software operations which are not sufficiently expressed in tests?
- Is there a need for extended tests, or if extended tests exists, have they been implemented in an appropriate way, and are they appropriately documented?

Test seem appropriate for me.

#### Visualisation (where appropriate)

- Do visualisations aid the primary purposes of statistical interpretation of results?
- Are there any aspects of visualisations which could risk statistical misinterpretation?

I ran the Visualisation example in plot.coevfit(). It plots marginal posteriors and chains for some parameters. First I have the issue that the model is not described in the package so it is difficult to connect the parameters. However, if the model is multivariate, then joint posteriors would be very helpful.
#### Package Design

- Is the package well designed for its intended purpose?
Yes
- In relation to **External Design:** Do exported functions and the relationships between them enable general usage of the package?
Yes
- In relation to **External Design:** Do exported functions best serve inter-operability with other packages?
Cannt see this.
- In relation to **Internal Design:** Are algorithms implemented appropriately in terms of aspects such as efficiency, flexibility, generality, and accuracy?
Algorithms are delegated to STAN. Package uses the phylo format, which is currently the best supported phylogenetic tree format in R.
- In relation to **Internal Design:** Could ranges of admissible input structures, or form(s) of output structures, be expanded to enhance inter-operability with other packages?
There is no real standard in the field for returning general PCM models so I do not see a need for this right now.

---

Estimated hours spent reviewing: 4h

- [ X ] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.
