### 2026-08-17 #717 @ScottClaessens
Thanks @krzbar for taking the time to review our package. I've updated our evolving to-do list associated with this submission (see issue [here](https://github.com/ScottClaessens/coevolve/issues/113)). I'll get to work on the revisions and post my response here soon.
