### 2026-08-24 #717 @ScottClaessens
I've finished my response to the reviews, but I'm just going to fix those final issues in the check above before submitting and recording it.
