### 2026-08-25 #717 @ScottClaessens
## Author response to `coevolve` package reviews

Thanks to both reviewers for taking the time to review the `coevolve` package. We really appreciate it!

I have made the requisite changes in version `1.2.0` of the package (see PR [here](https://github.com/ScottClaessens/coevolve/pull/117) and v1.2.0 release [here](https://github.com/ScottClaessens/coevolve/releases/tag/v1.2.0)). Below is a summary of the cumulative changes I've made in response to both reviewers.

### Reviewer 1

- Fixed the issue with `@srrstats {G3.0}`. `devtools::run_examples(document = FALSE)` now runs without error.
- Grouped package functions in the website reference.
- Set `patchwork` version >= 1.2.0 in Imports.
- Ensured that vignettes are up-to-date and include all error messages.
- Explained `elpd_diff` in the model comparison vignette as a way to compare models.
- Included `loo` as suggested package for model comparison in Suggests.
- Added @robmoss as a package reviewer ("rev" role) in the package DESCRIPTION file.

### Reviewer 2

- Added the stochastic evolutionary model equation to the README and the introductory vignette, as well as links to further resources for users to learn more about the model underlying the package.
- Amended the documentation for `coev_simulate_coevolution()` to clarify that the VAR(1) model underlying this function is not the same as the SDE model underlying the main package model.
- Created a new vignette about setting manual priors (see [here](https://scottclaessens.github.io/coevolve/articles/priors.html)).
- Added a method `pairs.coevfit` to produce multivariate pairs plots of joint posterior draws, as well as tests for this method.
- Added @krzbar as a package reviewer ("rev" role) in the package DESCRIPTION file.

I have responded in detail to Reviewer 1 in the comments above ([here](https://github.com/ropensci/software-review/issues/717#issuecomment-4982908564) and [here](https://github.com/ropensci/software-review/issues/717#issuecomment-4989877969)), so I won't repeat myself here. Regarding Reviewer 2's suggestions, we hope that the inclusion of the stochastic evolutionary model equation in the documentation, the inclusion of a vignette dedicated to setting priors, and the addition of a pairs method for plotting joint posterior draws addresses the issues raised by this reviewer.

All tests and checks pass. The one exception is the failed continuous integration check in `pkgcheck` above, which states that the package fails continuous integration checks. I'm not sure why this is happening as the CI checks all passed on GitHub (see [here](https://github.com/ScottClaessens/coevolve/actions/runs/32848288531)).
