### 2026-09-06 #717 @joelnitta
Hi @robmoss and @krzbar, thank you both again for your thorough reviews!

@ScottClaessens has posted a [response to the reviews](https://github.com/ropensci/software-review/issues/717#issuecomment-5411606279) and made the corresponding changes in [v1.2.0](https://github.com/ScottClaessens/coevolve/releases/tag/v1.2.0) (PR [#117](https://github.com/ScottClaessens/coevolve/pull/117); see also [NEWS.md](https://github.com/ScottClaessens/coevolve/blob/main/NEWS.md)).

Apologies for the delay in processing this, but we are now able to move to the next step: could @robmoss and @krzbar each take a look at the [response](https://github.com/ropensci/software-review/issues/717#issuecomment-5411606279) and the updated package and let me know whether your comments have been addressed? Please reply using the [reviewer approval template](https://devguide.ropensci.org/approval2template.html):

```
- [ ] The author has responded to my review and made changes to my satisfaction. I recommend approving this package.

Estimated hours spent reviewing:
```

(@robmoss please note that @ScottClaessens [responded to your review](https://github.com/ropensci/software-review/issues/717#issuecomment-4989877969) before the second review was finished, but I would appreciate your check at this point now that the package includes changes reflecting comments from both reviews)

We aim for about **one week** for this step, but no problem if you need a little longer — just let me know. Follow-up questions and further back-and-forth with the author are welcome here in the thread.

A couple of specific things it would help to hear your view on:

- @robmoss — whether these resolve the points from your review:
  - the new vignettes now show sampler warnings in the rendered output (see [Averaging over samples of trees](https://scottclaessens.github.io/coevolve/articles/multiphylo.html))
  - `elpd_diff` is explained in the new [Model comparison](https://scottclaessens.github.io/coevolve/articles/compare.html) vignette
  - `patchwork (>= 1.2.0)` is [pinned in Imports](https://github.com/ScottClaessens/coevolve/blob/3af7e1d2f5ed3d17a384bcbb5b24d6e063592e20/DESCRIPTION#L47)

- @krzbar — whether these resolve the points from your review:
  - the underlying stochastic model is now written out in the [README](https://github.com/ScottClaessens/coevolve#readme) and the [introductory vignette](https://scottclaessens.github.io/coevolve/articles/coevolve.html), with a link to the methods paper — your stated main reason for recommending silver
  - a dedicated ["Setting manual priors"](https://scottclaessens.github.io/coevolve/articles/priors.html) vignette, including its limitations (priors are set globally)
  - a new [`pairs()` method](https://scottclaessens.github.io/coevolve/reference/pairs.coevfit.html) for joint posterior draws, per your visualisation comment

  Given the model documentation, does your badge-grade recommendation change? And does your comment about algorithmic scaling not being discussed in the vignettes still stand?

Thanks!
